@@ -1,30 +1,20 @@
-import React from 'react';
-import { useQuery } from '@apollo/client';
+import React from 'react'
+// import { useQuery } from '@apollo/client';
 
-import ProfileList from '../components/ProfileList';
-
-import { QUERY_PROFILES } from '../utils/queries';
+// import component to be rendered
+import Jumbotron from '../components/Jumbotron';
+// import the queries if needed
 
 const Home = () => {
-  const { loading, data } = useQuery(QUERY_PROFILES);
-  const profiles = data?.profiles || [];
-
-  return (
-    <main>
-      <div className="flex-row justify-center">
-        <div className="col-12 col-md-10 my-3">
-          {loading ? (
-            <div>Loading...</div>
-          ) : (
-            <ProfileList
-              profiles={profiles}
-              title="My Bandwagon list"
-            />
-          )}
-        </div>
-      </div>
-    </main>
-  );
+    return (
+        <main>
+            <div className="flex-row justify-center">
+                <div className="col-12">
+                    <Jumbotron />
+                </div>
+            </div>
+        </main>
+    );
 };
 
 export default Home;
