@@ -1,66 +1,21 @@
-import React from 'react';
+import React from 'react'
+// import { useQuery } from '@apollo/client';
 
-import { Navigate, useParams } from 'react-router-dom';
-import { useQuery } from '@apollo/client';
-
-import SkillsList from '../components/SkillsList';
-import SkillForm from '../components/SkillForm';
-
-import { QUERY_SINGLE_PROFILE, QUERY_ME } from '../utils/queries';
-
-import Auth from '../utils/auth';
+// import components to be rendered
+// import Component from '../components/componentName';
+// import the queries if needed
 
 const Profile = () => {
-  const { profileId } = useParams();
-
-  // If there is no `profileId` in the URL as a parameter, execute the `QUERY_ME` query instead for the logged in user's information
-  const { loading, data } = useQuery(
-    profileId ? QUERY_SINGLE_PROFILE : QUERY_ME,
-    {
-      variables: { profileId: profileId },
-    }
-  );
-
-  // Check if data is returning from the `QUERY_ME` query, then the `QUERY_SINGLE_PROFILE` query
-  const profile = data?.me || data?.profile || {};
-
-  // Use React Router's `<Redirect />` component to redirect to personal profile page if username is yours
-  if (Auth.loggedIn() && Auth.getProfile().data._id === profileId) {
-    return <Navigate to="/me" />;
-  }
-
-  if (loading) {
-    return <div>Loading...</div>;
-  }
-
-  if (!profile?.name) {
     return (
-      <h4>
-        You need to be logged in to see your profile page. Use the navigation
-        links above to sign up or log in!
-      </h4>
+        <main>
+            <div className="flex-row justify-center">
+                <div className="col-12 col-md-10 my-3">
+                    This is the profile page
+                    {/* <Component /> */}
+                </div>
+            </div>
+        </main>
     );
-  }
-
-  return (
-    <div>
-      <h2 className="card-header">
-        {profileId ? `${profile.name}'s` : 'Your'} friends have endorsed these
-        skills...
-      </h2>
-
-      {profile.skills?.length > 0 && (
-        <SkillsList
-          skills={profile.skills}
-          isLoggedInUser={!profileId && true}
-        />
-      )}
-
-      <div className="my-4 p-4" style={{ border: '1px dotted #1a1a1a' }}>
-        <SkillForm profileId={profile._id} />
-      </div>
-    </div>
-  );
 };
 
 export default Profile;
